Fix e2e timeout
