+++ test/diamond-e2e/testdata/cycle-to-self.yaml
