+++ test/diamond-e2e/testdata/join-on-sink.yaml
