+++ test/transformer-e2e/testdata/source-filtering.yaml
